refactor: redis repository를 qna로 옮김
