+++ apps/qna/redis/__init__.py
@@ -1,5 +1,7 @@
 from .factories import CacheFactory
+from .repositories import CacheRepository
 
 __all__ = [
     "CacheFactory",
+    "CacheRepository",
 ]
