+++ apps/qna/redis/tests/test_repositories.py
@@ -7,7 +7,7 @@ from apps.core.utils.isolated_cache_testcase import (
     FixedPrefixRedisTestClient,
     IsolatedRedisTestClient,
 )
-from apps.core.utils.redis_repository import CacheRepository
+from apps.qna.redis import CacheRepository
 from apps.qna.dtos import InitialQNA, Message
